Move converter registration tests
